client: Implement delayed reconnection logic.
If we EOF, then delay a little. If we're KILLed or banned or something, wait a
while longer to reconnect. If we fail to connect to the server, also, wait a
while longer as the problem is probably not a transient one.

Thanks to John for the tip about using timers.