gerrit: Reset reconnection delay on successful stream-events.
We can't wait for the first line to be read out, because in the case where no
such line is read for a while (e.g. overnight) we'll get multiple disconnects
before someone does something on gerrit thanks to the gerrit sshd's disconnect
behaviour, which leads to multi-second reconnect delay, and possibly missed
events.