client: Drop WriteDeadline, reset ReadDeadline each read.
The design for deadlines seems to be a little braindead, as they require an
absolute time rather than a timespan (thus: you need to reset them every time
you read or write).

Doing this for write would be too burdensome, so we rely on the server pinging
us as our keepalive mechanism. We won't instantly discover a disconnection, but
this is probably good enough.