client: Send PING every minute (from client to server)...
...rather than just relying on server PINGing us.

In the case where the client connection is very active (sending at least one
message every minute or so), on some IRC servers, we won't be guarenteed to
recieve a PING (as they only PING inactive connections).

This was causing us to trip over the ReadDeadline, and disconnect from IRC at
regular intervals; visible with qt_gerrit disconnecting every so often during
the European workday (when #qt-gerrit is full of announcements).