gerrit: Don't reuse the channel
Picture the case where we have a line pending read (r1), but about to time
out (t0). t0 times out, the conn is closed.

The parent goroutine connects, sets up a new reader goroutine (r2), and waits
again. However, in the meantime, we read an EOF from r1, and close the
connection, which then proceeds to continue on cascading to r2 etc.