Benchmark and performance improvements
This takes Detect() down from 6 allocations/op to one allocation per op
(in xbuffer), and gives a total performance win on jpeg size detection
of around 15%.

One small API cost is that ImageSize structs may no longer be nil in the error
case, but this is not too significant a problem.

Another small note is that detector instances are no longer
concurrent-safe: due to internal buffer sharing, you should not share a detector
between goroutines.

The implementation also formalizes the dependency on xbuffer, as no
detector will function without it (due to casts).