Re-implement exact tag matches
Use the syntax: ^tag$ to search for an exact match for 'tag'.
This also will allow us to implement partial matching (at begin or end)
later, though we might need to do some more work on the Index interface
to allow that to make sense (rather than using QueryTagFuzzy all the
time).