Add some (admittedly convoluted) testing
For each test, we encode a message with some expected values.
We check the message encodes.
We then decode the bytes, and check it decodes.

Then we take the bytes, build a Go protobuf stub to decode the same
bytes, and make sure it *also* gets the same expected end result.

This way we ensure that we can round-trip with ourselves, but also that
Go can round-trip with us.