Expand EXIF parsing to handle more things
This also breaks API: exif data is now moved from ImageInfo to its own
struct, to make it clear what data comes from what source.

Performance might also be a tiny bit better due to not reading the image
into a newly constructed buffer (instead, using ReadSlice) to sub-slice
from the data that is already read, but I haven't measured that
scientifically.

This is still incomplete; there's plenty of EXIF tags that are not
fetched, and date information might do with a custom type to allow for
easily extracting components out of it (dd/mm/yy etc), but it's a start.

Orientation parsing also appears to be wrong on some images, I suspect
due to endianness problems, but that's a problem for another day, too.